moved sigmaM
